+++ packages/dev/core/src/Materials/effect.ts
@@ -10,12 +10,11 @@ import { ShaderProcessor } from "../Engines/Processors/shaderProcessor";
 import type { IShaderProcessor } from "../Engines/Processors/iShaderProcessor";
 import type { ProcessingOptions, ShaderCustomProcessingFunction, ShaderProcessingContext } from "../Engines/Processors/shaderProcessingOptions";
 import type { IMatrixLike, IVector2Like, IVector3Like, IVector4Like, IColor3Like, IColor4Like, IQuaternionLike } from "../Maths/math.like";
-import type { ThinEngine } from "../Engines/thinEngine";
+import type { AbstractEngine } from "../Engines/abstractEngine";
 import type { IEffectFallbacks } from "./iEffectFallbacks";
 import { ShaderStore as EngineShaderStore } from "../Engines/shaderStore";
 import { ShaderLanguage } from "./shaderLanguage";
 
-import type { Engine } from "../Engines/engine";
 import type { InternalTexture } from "../Materials/Textures/internalTexture";
 import type { ThinTexture } from "../Materials/Textures/thinTexture";
 import type { RenderTargetTexture } from "../Materials/Textures/renderTargetTexture";
@@ -201,7 +200,7 @@ export class Effect implements IDisposable {
 
     private static _UniqueIdSeed = 0;
     /** @internal */
-    public _engine: Engine;
+    public _engine: AbstractEngine;
     private _uniformBuffersNamesList: string[];
     private _uniformsNames: string[];
     private _samplers: { [key: string]: number } = {};
@@ -275,9 +274,9 @@ export class Effect implements IDisposable {
     constructor(
         baseName: IShaderPath | string,
         attributesNamesOrOptions: string[] | IEffectCreationOptions,
-        uniformsNamesOrEngine: string[] | ThinEngine,
+        uniformsNamesOrEngine: string[] | AbstractEngine,
         samplers: Nullable<string[]> = null,
-        engine?: ThinEngine,
+        engine?: AbstractEngine,
         defines: Nullable<string> = null,
         fallbacks: Nullable<IEffectFallbacks> = null,
         onCompiled: Nullable<(effect: Effect) => void> = null,
@@ -291,7 +290,7 @@ export class Effect implements IDisposable {
 
         if ((<IEffectCreationOptions>attributesNamesOrOptions).attributes) {
             const options = <IEffectCreationOptions>attributesNamesOrOptions;
-            this._engine = <Engine>uniformsNamesOrEngine;
+            this._engine = <AbstractEngine>uniformsNamesOrEngine;
 
             this._attributesNames = options.attributes;
             this._uniformsNames = options.uniformsNames.concat(options.samplers);
@@ -315,7 +314,7 @@ export class Effect implements IDisposable {
             this._processFinalCode = options.processFinalCode ?? null;
             this._processCodeAfterIncludes = options.processCodeAfterIncludes ?? undefined;
         } else {
-            this._engine = <Engine>engine;
+            this._engine = <AbstractEngine>engine;
             this.defines = defines == null ? "" : defines;
             this._uniformsNames = (<string[]>uniformsNamesOrEngine).concat(<string[]>samplers);
             this._samplerList = samplers ? <string[]>samplers.slice() : [];
@@ -475,7 +474,7 @@ export class Effect implements IDisposable {
      * The engine the effect was initialized with.
      * @returns the engine.
      */
-    public getEngine(): Engine {
+    public getEngine(): AbstractEngine {
         return this._engine;
     }
