+++ packages/dev/core/src/XR/features/WebXRLightEstimation.ts
@@ -14,6 +14,7 @@ import { BaseTexture } from "../../Materials/Textures/baseTexture";
 import { SphericalHarmonics, SphericalPolynomial } from "../../Maths/sphericalPolynomial";
 import { LightConstants } from "../../Lights/lightConstants";
 import { HDRFiltering } from "core/Materials/Textures/Filtering/hdrFiltering";
+import type { ThinEngine } from "core/Engines";
 
 /**
  * Options for Light Estimation feature
@@ -159,7 +160,7 @@ export class WebXRLightEstimation extends WebXRAbstractFeature {
             this.directionalLight.falloffType = LightConstants.FALLOFF_GLTF;
         }
 
-        this._hdrFilter = new HDRFiltering(this._xrSessionManager.scene.getEngine());
+        this._hdrFilter = new HDRFiltering(this._xrSessionManager.scene.getEngine() as ThinEngine);
 
         // https://immersive-web.github.io/lighting-estimation/
         Tools.Warn("light-estimation is an experimental and unstable feature.");
@@ -191,7 +192,7 @@ export class WebXRLightEstimation extends WebXRAbstractFeature {
 
     private _getCanvasContext(): WebGLRenderingContext | WebGL2RenderingContext {
         if (this._canvasContext === null) {
-            this._canvasContext = this._xrSessionManager.scene.getEngine()._gl;
+            this._canvasContext = (this._xrSessionManager.scene.getEngine() as ThinEngine)._gl;
         }
         return this._canvasContext;
     }
